Osc and FX code cleanup; Version to 180
- Oscillators are properly named
- Oscillators and FX get distinct .h and .cpp per
- Version moves to 1.8

Closes #2605
Closes #2679
