+++ src/common/dsp/SampleAndHoldOscillator.cpp
@@ -13,8 +13,9 @@
 ** open source in September 2018.
 */
 
-#include "Oscillator.h"
+#include "SampleAndHoldOscillator.h"
 #include "DspUtilities.h"
+
 
 using namespace std;
