+++ src/common/dsp/VectorizedSvfFilter.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <vt_dsp/portable_intrinsics.h>
 
 //------------------------------------------------------------------------------------------------
